Fix: Add common error class
